Fix Docker build & start
Updates the Docker configuration to streamline the deployment process for the Voyagers of Nera dedicated server.

Updates the server update logic to update only if older than 14 days.

Migrates to a steamcmd-based image for simpler server setup.

Also, adjusts image name and labels in the docker-compose configuration for clarity.

Still need to publish the image & test the docker compose